refactor: extract single-name marketplace edge handler factory
info.ts and remove.ts were identical except for their USAGE string and the
orchestrator they delegate to (getMarketplaceInfo vs removeMarketplace).
Extract a new makeSingleNameMarketplaceHandler factory in marketplace/
shared.ts that owns the single-name-positional parseCommandArgs parse, the
Missing-required-argument error callback, the undefined-guard, and the
delegate call. Both orchestrator option types structurally satisfy the run
param (RemoveMarketplaceOptions's optional cascade? is omittable).

Rewrite both handlers as thin wrappers; each keeps its own USAGE constant.
Byte-neutral: notified strings and USAGE values unchanged; catalog-uat
byte gate green.
