+++ extensions/pi-claude-marketplace/edge/handlers/marketplace/info.ts
@@ -8,8 +8,8 @@
 // shape and delegates.
 
 import { getMarketplaceInfo } from "../../../orchestrators/marketplace/info.ts";
-import { notifyUsageError } from "../../../shared/notify.ts";
-import { parseCommandArgs } from "../../args-schema.ts";
+
+import { makeSingleNameMarketplaceHandler } from "./shared.ts";
 
 import type { ExtensionAPI, ExtensionCommandContext } from "../../../platform/pi-api.ts";
 
@@ -18,30 +18,5 @@ const USAGE = "Usage: /claude:plugin marketplace info <name> [--scope user|proje
 export function makeMarketplaceInfoHandler(
   pi: ExtensionAPI,
 ): (args: string, ctx: ExtensionCommandContext) => Promise<void> {
-  return async (args, ctx): Promise<void> => {
-    const parsed = parseCommandArgs(
-      args,
-      {
-        positional: [{ name: "name" }] as const,
-        usage: USAGE,
-      },
-      (message) => {
-        notifyUsageError(ctx, {
-          message: message === USAGE ? "Missing required argument." : message,
-          usage: USAGE,
-        });
-      },
-    );
-    if (parsed === undefined) {
-      return;
-    }
-
-    await getMarketplaceInfo({
-      ctx,
-      pi,
-      name: parsed.name,
-      cwd: ctx.cwd,
-      ...(parsed.scope !== undefined && { scope: parsed.scope }),
-    });
-  };
+  return makeSingleNameMarketplaceHandler(pi, USAGE, getMarketplaceInfo);
 }
