+++ extensions/pi-claude-marketplace/edge/handlers/marketplace/remove.ts
@@ -14,8 +14,8 @@
 // positional/scope shape.
 
 import { removeMarketplace } from "../../../orchestrators/marketplace/remove.ts";
-import { notifyUsageError } from "../../../shared/notify.ts";
-import { parseCommandArgs } from "../../args-schema.ts";
+
+import { makeSingleNameMarketplaceHandler } from "./shared.ts";
 
 import type { ExtensionAPI, ExtensionCommandContext } from "../../../platform/pi-api.ts";
 
@@ -24,30 +24,5 @@ const USAGE = "Usage: /claude:plugin marketplace <remove|rm> <name> [--scope use
 export function makeRemoveHandler(
   pi: ExtensionAPI,
 ): (args: string, ctx: ExtensionCommandContext) => Promise<void> {
-  return async (args, ctx): Promise<void> => {
-    const parsed = parseCommandArgs(
-      args,
-      {
-        positional: [{ name: "name" }] as const,
-        usage: USAGE,
-      },
-      (message) => {
-        notifyUsageError(ctx, {
-          message: message === USAGE ? "Missing required argument." : message,
-          usage: USAGE,
-        });
-      },
-    );
-    if (parsed === undefined) {
-      return;
-    }
-
-    await removeMarketplace({
-      ctx,
-      pi,
-      name: parsed.name,
-      cwd: ctx.cwd,
-      ...(parsed.scope !== undefined && { scope: parsed.scope }),
-    });
-  };
+  return makeSingleNameMarketplaceHandler(pi, USAGE, removeMarketplace);
 }
